airflow/requirements.txt
apache-airflow-providers-cncf-kubernetes
psycopg2-binary